+++ backend/etl/leads/scrape.py
@@ -7,10 +7,10 @@ from selenium.common.exceptions import (
 )
 from selenium.webdriver.remote.webelement import WebElement
 
-from app.core.chrome import Driver
 from app.core.conf import linkedin, settings
 from app.etl.base import Job
 from app.logging import console_log
+from backend.etl.chrome import Driver
 
 
 async def search_results():
